#37 refactor: organize strategy package
