+++ src/main/java/br/ufpr/webII/trabalhoFinal/infra/service/strategy/ValidateStatusChangeContext.java
@@ -1,4 +1,4 @@
-package br.ufpr.webII.trabalhoFinal.infra.service;
+package br.ufpr.webII.trabalhoFinal.infra.service.strategy;
 
 import br.ufpr.webII.trabalhoFinal.domain.request.RequestUpdateDTO;
